Added images, score, camera follow and packet optimization
Added my own art to the files and included them in the game to show visual information like players and bullets,

Added a score to the top left, this is found using new variables inside of the player class id and score,

Added script to allow the camera to follow where the player moves to, making it seem like the map moves rather than the player,

Bullets at the moment are not facing the correct direction when firing, code needs adjustment to rotate based on mouse angle.

Changed how data is sent, so that data that has changed is sent, rather than all of the data. This allows for a more stable and higher performance game. This is crucial when making multiplayer online games.