Introduced map system
I have added a second map which is more or less the same as the original map,

I have introduced parameters in the constructors of entity, player and bullet objects,

Changed render settings to only render objects on the same map as the player/s.This allows for both maps (maybe more later) to exist on the same webpage, essentially having two maps in the same space that can't interact with each other in anyway, this also holds true for any objects (like players and bullets) on those maps, they cannot interact with any other object that isn't on the same map,

After creating the Heroku application I am no longer able to test locally before pushing any updates to the live server. I have tried to find the answer to why this is, however research into this has been inconclusive. It is likely it could be to do with using UTF-16 line endings as windows 10 takes UTF-8 from what I have read (I haven't looked into how I would change this, but is linked to having created a heroku server for the project) and will not accept the UTF-16 syntax. Otherwise this is likely just a silly syntax error, either way I will not find out until I have pushed the update up to Github and into Heroku.

*Update* After quickly inspecting syntaxes inside of app.js on the change log of Github I was able to find 1 syntax error and 2 reference errors that was causing the local hosting issue. This is now fixed and local hosting is now available to me again, map system has been integrated successfully and is working perfectly, with no known bugs at this time.