Change map functionality and Client profiling - attempted server profiling
Changed html layout to allow for more customisability between different elements, i.e. adding a button over the top of the game, whilst still having a separate layer for drawing the score on top of the game and a separate layer for the game itself,

Added button that allows the user to change the map they are on, future updates may add more functionality once more maps are added,

Made button text and chat text unselectable in chrome, support for others maybe in future but no need for now as testing is occuring on chrome, as chrome is the most popular browser,

started profiling on client side using chrome dev tools (performance tab) which was used to check how exhaustive the code was and to show which parts were more exhaustive than others. I learnt that the drawScore and shootBullet functions were very exhaustive.

I changed the drawScore function so that it was only updated if it had changed, this drastically reduced its load on the browser,

I changed the shooting mechanic so that only one bullet could be fired every half second. Unfortunately I was unable to do this while allowing the player to hold the mouse button to keep shooting every half second. Creating a semi automatic gun effect rather than a automatic/continuously firing gun effect,

I also attempted to start profiling server side but was unable to npm install v8-profiler, due to errors on install, and could not find a suitable replacement for this. Attempts will be made in future to either install the module or otherwise find a suitable replacement.