Created entity.js to start the modularisation process
Put entity, player and bullet classes into a separate file called entity.js,

Made these classes global along with SOCKET_LIST,

Changed code used to initialise, remove and otherwise update player information between server and client to fit the new global classes in the new file,

Code was made more modular however this was a quick and easy way to achieve this and is not very secure and definitely not the best of practises. However this ensures that I do not have to use other libraries like exports or handlebars. On this short project I feel that it would be best to spend time making the mechanics stronger rather than learning to use another library, and keep this as a stretch goal.