60 fps, Survival Time, score submitted to leaderboard database
Changed intervals to allow for 60 frames per second. After doing this the movement of entity's became too fast (bullets and players) so I changed the pace of these to half of what they were originally which solved the issue and allowed for a much more precise rendering of objects that looks smoother to users,

Added a survival time variable to allow the player to know how long they survived this can easily be integrated into the leaderboard in some manner later on,

Added functions that submit the users score to a new database collection upon death. This took a very long time to get right as I have never used databases before and as such have never used or understood the method calls before now. After much research through stack overflow and mongoDB documentation I finally created a function that could update the score attribute in the new collection. However when testing I found that this function did not create new documents in the collection when a new user (who had not died before) died and so the game would crash. I fixed this by making the addUser function insert the players username and a score of 0 when they create their login details.

This function uses a global variable to be allowed to use the sockets in app.js. When adding in handlebars or another file linking library I will be able to access this without the need for a global variable, but I deem this unimportant at the present moment due to time constraints.