removal of unnecessary code, Replaced some images, Fixed minor mistakes in code, Added messages to chat to inform players of how long a player who was killed survived for
Some code throughout files removed as is uneccesary,

Replaced the bullet and player images,

Bullet image replaced as it is very costly to rotate images to make the bullet look like it is facing the direction it is travelling,

Replaced player image to make dimensions more easily manageable by hand now 1000x1000 rather than 10003x1198,

Some mistakes that have been left have been fixed like ctx -> ctxUI with the font change,

Added a system that will, on a players death, alert the player in chat that they have survived for so long, whilst telling everyone else which player has died and how long they survived for.

Some work needs to be done on the globalSocketCall that calls IsScoreHigher as I have only just realised that I am able to call socket.emits and .ons from within entity.js (before I thought I would have to redeclare a socket to be able to do this and didn't see I had placed some of the socket code in with Entity).