isScoreHigher optimisation
Moved isScoreHigher into Entity.js for now so that I don't have to use a global variable to call the method isScoreHigher,

This does mean that I have to make the database variable global however, at least until I add a link via handlebars, etc.

Added a call for this method when the player disconnects so that if they disconnect whilst having a high score of theirs, this will still be added to the database, and they will not have to wait to die before they will achieve a new high score,

After implementing this there were some issues with the localhosting whenever I would refresh an old localhost page. This would cause an error because no user was logged in as the session had expired and the code was not able to identify p.username. I added some code to set the user to guest and score to 0 so that this would not cause a server crash.

Changed the bullet timer to 250 so that bullets will now not be deleted for roughly 12.5 seconds giving them more time to travel across the large map, probably only need about 8 seconds.

Some obsolete and commented code has been deleted as it will not be used any longer