Leaderboard fully implemented, respawn changed due to leaderboard
Added new div container to contain leaderboard, keeping this hidden whilst the user is loggingin, signing up and playing the game. This is then only shown when the player dies.

Added code to respawn the player at co-ord's 10000,10000 on death so when they are looking at the leaderboard they cannot be killed or shoot,

Added code to grab the leaderboard database collection from the online database and sort it by score (descending) to display upon the new div container showing when the player is dead,

Whilst the code added to ensure the player does not die whilst looking at the leaderboard is a bit hacky, it works for now which is the main aim. This will be made better in the future after modularisation of the code.