Added pickups, Added images for pickups, Added button for changing map whilst dead, Added code to allow pickups to work (Health at least...)
I added a button on the leaderboard screen for users to change the map before they respawn, the one on the game screen is planned to be removed before "release",

Created and added images for use with health and score pickups,

Added constructors and code for the two added pickups so that they can be picked up by the player and health or score added to the specific player who picked up the said pickups,

Unfortunately the score pickups are not currently spawning due to what I can only presume is a huge error on my part where I delete the list somehow before they are rendered. I cannot find this problematic code anywhere after looking for hours.  When using console.logs to see the full list it seems that when a player is connected they seem to be deleted almost instantly. This means that the problem is likely to be inside some code pertaining to the player, i.e. inside the players methods that are called constantly. I have looked but have still not found anything so hopefully this is resolved soon.